organize functions, update to node 22
