+++ svelte.config.js
@@ -12,7 +12,7 @@ const config = {
 		// If your environment is not supported or you settled on a specific environment, switch out the adapter.
 		// See https://kit.svelte.dev/docs/adapters for more information about adapters.
 		adapter: adapterVercel({
-			runtime: 'nodejs20.x',
+			runtime: 'nodejs22.x',
 		}),
 	},
 };
